+++ webview-ui/src/components/ui/checkbox-group/CheckboxGroup.vue
