--- webview-ui/src/components/ui/popover/index.ts
@@ -1,3 +1,0 @@
-export { default as Popover } from './Popover.vue'
-export { default as PopoverTrigger } from './PopoverTrigger.vue'
-export { default as PopoverContent } from './PopoverContent.vue'
